some more modelling of system calls and parameters
some prototypic code to track open file descriptions. However, this seems
somewhat unnecessary as I noticed in the end, because the procfs provides us
with this information from the kernel (and he knows best, doesn't he?)

Finally a first bugfix / workaround for the case when exceptions are thrown
during trace handling: the destructor of the tracedproc class tried to wait on
the child but the child was stuck due to the trace operation. So we send the
child a sigkill in this situation.