autotools: fix in-tree build
Turned out that we attempted to create an executable "wmfs" where a directory
"wmfs" was already in place. Thus only out-of-tree builds worked.

Renaming the executable would have been stupied. Restructuring the makefiles
to output the executable in a different directory would also have been
annoying. So I've renamed the directory to main instead.