AbortHandler: factored out blocking call abortion logic into new class
To support the concept of aborting a blocking call more broadly than
just in the EventFile we now have a mixin class AbortHandler which can
be instantiated via the Entry class to get the behaviour without having
an overly specialized and deep class hierarchy.

This code will be needed in an upcoming commit for the selection buffer
access.