global: use libxpp API instead of local classes from src/x11
Similarly to what was done for libcosmos this change replaces now
redundant logic mostly found in src/x11 by the API found in libxpp.

This is quite an intrusive change. Some of the Xwmfs specific logic was
dropped from libxpp, thus we need a new WinManagerWindow which derives
from xpp::RootWin to re-add the Xwmfs specific logic we need.

This change also made some additions in libxpp itself necessary. The
resulting program still works as before, as far as I could test it for
the moment.