Makefile.am: add missing sources and use paths without symlinks
Autotools really are as stubborn as an old ox. There's just no way to
add a symlink to the distribution tarball and not have it dereference
when tarring it.

Therefore we need to references libcosmos and libxpp sources including
the symlink, so that when it is dereferences everything is in the place
expect by the sources.

Also add a couple of missing headers.